fix: better testing to handle different default branches
