+++ pkg/keeper/keeper_test.go
@@ -29,8 +29,7 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/jenkins-x/lighthouse/pkg/git/v2"
-	"github.com/stretchr/testify/require"
+	"github.com/jenkins-x/lighthouse/pkg/gittest"
 
 	"github.com/jenkins-x/go-scm/scm"
 	fakescm "github.com/jenkins-x/go-scm/scm/driver/fake"
@@ -1025,27 +1024,8 @@ func TestCheckMergeLabels(t *testing.T) {
 	}
 }
 
-func GetDefaultBranch(t *testing.T) string {
-	logger := logrus.WithField("client", "git")
-	censor := func(content []byte) []byte { return content }
-
-	executor, err := git.NewCensoringExecutor(".", censor, logger)
-	require.NoError(t, err, "failed to find git binary")
-
-	defaultBranch := "master"
-	out, err := executor.Run("config", "--global", "--get", "init.defaultBranch")
-	if err == nil {
-		text := strings.TrimSpace(string(out))
-		if text != "" {
-			defaultBranch = text
-		}
-	}
-	t.Logf("using default branch: %s\n", defaultBranch)
-	return defaultBranch
-}
-
 func TestTakeAction(t *testing.T) {
-	defaultBranch := GetDefaultBranch(t)
+	defaultBranch := gittest.GetDefaultBranch(t)
 
 	sleep = func(time.Duration) {}
 	defer func() { sleep = time.Sleep }()
