+++ pkg/repoowners/repoowners_test.go
@@ -24,6 +24,8 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/jenkins-x/lighthouse/pkg/gittest"
+
 	"github.com/jenkins-x/lighthouse/pkg/config/lighthouse"
 	"github.com/jenkins-x/lighthouse/pkg/git/localgit"
 	"github.com/jenkins-x/lighthouse/pkg/scmprovider/fake"
@@ -102,6 +104,7 @@ func patternAll(values ...string) map[string]sets.String {
 }
 
 func getTestClient(
+	defaultBranch string,
 	files map[string][]byte,
 	enableMdYaml,
 	skipCollab,
@@ -140,7 +143,7 @@ func getTestClient(
 				}
 			}
 		}
-		if err := localGit.Checkout("org", "repo", "master"); err != nil {
+		if err := localGit.Checkout("org", "repo", defaultBranch); err != nil {
 			return nil, nil, err
 		}
 	}
@@ -175,6 +178,8 @@ func getTestClient(
 }
 
 func TestOwnersDirExcludes(t *testing.T) {
+	defaultBranch := gittest.GetDefaultBranch(t)
+
 	validatorExcluded := func(t *testing.T, ro *RepoOwners) {
 		for dir := range ro.approvers {
 			if strings.Contains(dir, "src") {
@@ -213,13 +218,13 @@ func TestOwnersDirExcludes(t *testing.T) {
 	}
 
 	getRepoOwnersWithExcludes := func(t *testing.T, defaults []string, byRepo map[string][]string) *RepoOwners {
-		client, cleanup, err := getTestClient(testFiles, true, false, true, defaults, byRepo, nil)
+		client, cleanup, err := getTestClient(gittest.GetDefaultBranch(t), testFiles, true, false, true, defaults, byRepo, nil)
 		if err != nil {
 			t.Fatalf("Error creating test client: %v.", err)
 		}
 		defer cleanup()
 
-		ro, err := client.LoadRepoOwners("org", "repo", "master")
+		ro, err := client.LoadRepoOwners("org", "repo", defaultBranch)
 		if err != nil {
 			t.Fatalf("Unexpected error loading RepoOwners: %v.", err)
 		}
@@ -271,6 +276,8 @@ func TestOwnersDirExcludes(t *testing.T) {
 }
 
 func TestOwnersRegexpFiltering(t *testing.T) {
+	defaultBranch := gittest.GetDefaultBranch(t)
+
 	tests := map[string]sets.String{
 		"re/a/go.go":   sets.NewString("re/all", "re/go", "re/go-in-a"),
 		"re/a/md.md":   sets.NewString("re/all", "re/md-in-a"),
@@ -280,13 +287,13 @@ func TestOwnersRegexpFiltering(t *testing.T) {
 		"re/b/md.md":   sets.NewString("re/all"),
 	}
 
-	client, cleanup, err := getTestClient(testFilesRe, true, false, true, nil, nil, nil)
+	client, cleanup, err := getTestClient(gittest.GetDefaultBranch(t), testFilesRe, true, false, true, nil, nil, nil)
 	if err != nil {
 		t.Fatalf("Error creating test client: %v.", err)
 	}
 	defer cleanup()
 
-	r, err := client.LoadRepoOwners("org", "repo", "master")
+	r, err := client.LoadRepoOwners("org", "repo", defaultBranch)
 	if err != nil {
 		t.Fatalf("Unexpected error loading RepoOwners: %v.", err)
 	}
@@ -304,6 +311,8 @@ func strP(str string) *string {
 }
 
 func TestLoadRepoOwners(t *testing.T) {
+	defaultBranch := gittest.GetDefaultBranch(t)
+
 	tests := []struct {
 		name              string
 		mdEnabled         bool
@@ -435,7 +444,7 @@ func TestLoadRepoOwners(t *testing.T) {
 		},
 		{
 			name:   "OWNERS from master branch while release branch diverges",
-			branch: strP("master"),
+			branch: strP(defaultBranch),
 			extraBranchesAndFiles: map[string]map[string][]byte{
 				"release-1.10": {
 					"src/doc/OWNERS": []byte("approvers:\n - maggie\n"),
@@ -496,14 +505,14 @@ func TestLoadRepoOwners(t *testing.T) {
 
 	for _, test := range tests {
 		t.Logf("Running scenario %q", test.name)
-		client, cleanup, err := getTestClient(testFiles, test.mdEnabled, test.skipCollaborators, test.aliasesFileExists, nil, nil, test.extraBranchesAndFiles)
+		client, cleanup, err := getTestClient(defaultBranch, testFiles, test.mdEnabled, test.skipCollaborators, test.aliasesFileExists, nil, nil, test.extraBranchesAndFiles)
 		if err != nil {
 			t.Errorf("Error creating test client: %v.", err)
 			continue
 		}
 		defer cleanup()
 
-		base := "master"
+		base := defaultBranch
 		if test.branch != nil {
 			base = *test.branch
 		}
@@ -554,6 +563,8 @@ func TestLoadRepoOwners(t *testing.T) {
 }
 
 func TestLoadRepoAliases(t *testing.T) {
+	defaultBranch := gittest.GetDefaultBranch(t)
+
 	tests := []struct {
 		name string
 
@@ -594,13 +605,13 @@ func TestLoadRepoAliases(t *testing.T) {
 		},
 	}
 	for _, test := range tests {
-		client, cleanup, err := getTestClient(testFiles, false, false, test.aliasFileExists, nil, nil, test.extraBranchesAndFiles)
+		client, cleanup, err := getTestClient(gittest.GetDefaultBranch(t), testFiles, false, false, test.aliasFileExists, nil, nil, test.extraBranchesAndFiles)
 		if err != nil {
 			t.Errorf("[%s] Error creating test client: %v.", test.name, err)
 			continue
 		}
 
-		branch := "master"
+		branch := defaultBranch
 		if test.branch != nil {
 			branch = *test.branch
 		}
